Commit
The goal of this project is to simulate an infinite underwater world as realistically as possible and to explore this world using a Submarin. The terrain is generated as the Submarin explores the world and it can be altered by a torpedo launched by the Submarin. At he same time multiple fishes will swim around the Submarin and will react to the surrounding environment and to the nearby fish. The project was developed using the C# programming language and the Unity Game Engine.